Scaffolding out classes
This creates several new modules and classes showing the basic structure of the class hierarchy, creating two teams of beetles with one beetle with the peashooter attack.

The classes are fully wired up for handling behavior in `on_update` and drawing on `on_draw`

This required a few changes to the design doc to match changes I made while creating the classes.